+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - [Fix #119] Trigger completer with the current text by default.
 - [Fix #120] Hide dropdown before destroying it.
 - [Fix #121] Don't throw an exception even if a jquery click event is manually triggered.
+- Add `match` as the third argument of a search function.
 
 ## `v0.3.2` - Sep 16, 2014